explicitly set cross compile vars for SCT
The SCT does not use GCC5_ARM_PREFIX, and thus SCT builds fall back
to defaults based on what is found in the $PATH.

For 32-bit SCT the default is arm-linux-gnueabihf- and unless you
have that compiler in your $PATH the build fails.

Set the SCT cross compile variables CROSS_COMPILE_64 and
CROSS_COMPILE_32 so the settings are explicit and not defaults.

Signed-off-by: Stuart Yoder <stuart.yoder@arm.com>
Signed-off-by: Grant Likely <grant.likely@arm.com>